requirements-dev.txt
astropy
astroquery
black
defusedxml
flake8
flake8-docstrings
flake8-print
isort
matplotlib
mypy
myst-parser
numpydoc
pikepdf
pytest
pytest-cov
pytest-regressions
python-dateutil
requests
responses
sphinx
sphinx-autodoc-typehints
sphinx-book-theme
tox
types-requests
